+++ packages/@vuetify/presets/generator/templates/base/src/layouts/default/AppBar.vue
@@ -1,6 +1,9 @@
 <template>
   <v-app-bar app>
-    <v-app-bar-nav-icon @click="drawer = !drawer" />
+    <v-app-bar-nav-icon
+      aria-label="Toggle Navigation Drawer"
+      @click="drawer = !drawer"
+    />
 
     <v-toolbar-title>Application</v-toolbar-title>
   </v-app-bar>
@@ -8,7 +11,7 @@
 
 <script>
   // Utilities
-  import { sync  } from 'vuex-pathify'
+  import { sync } from 'vuex-pathify'
 
   export default {
     name: 'DefaultAppBar',