+++ packages/@vuetify/presets/generator/templates/base/src/plugins/webfontloader.js
@@ -11,4 +11,8 @@ WebFontLoader.load({
   google: {
     families: ['Roboto:100,300,400,500,700,900&display=swap'],
   },
+  custom: {
+    urls: ['https://cdn.jsdelivr.net/npm/@mdi/font@5.x/css/materialdesignicons.min.css'],
+    timeout: 2000,
+  },
 })