+++ packages/vue-cli-plugin-vuetify/generator/templates/v3/src/App.js.vue
@@ -1,0 +1,31 @@
+<template>
+  <v-app>
+    <v-main>
+      <%_ if (router) { _%>
+      <router-view/>
+      <%_ } else { _%>
+      <HelloWorld/>
+      <%_ } _%>
+    </v-main>
+  </v-app>
+</template>
+
+<script>
+<%_ if (!router) { _%>
+import HelloWorld from './components/HelloWorld.vue';
+<%_ } _%>
+
+export default {
+  name: 'App',
+
+  <%_ if (!router) { _%>
+  components: {
+    HelloWorld,
+  },
+
+  <%_ } _%>
+  data: () => ({
+    //
+  }),
+};
+</script>