+++ packages/vue-cli-plugin-vuetify/generator/templates/v3/src/views/Home.js.vue
@@ -1,0 +1,17 @@
+<template>
+  <div>
+    <HelloWorld/>
+  </div>
+</template>
+
+<script>
+import HelloWorld from '../components/HelloWorld.vue';
+
+export default {
+  name: 'Home',
+  
+  components: {
+    HelloWorld,
+  },
+};
+</script>