+++ packages/vue-cli-plugin-vuetify/generator/tools/vuetify.js
@@ -1,27 +1,27 @@
 // Imports
-const fs = require('fs')
 const helpers = require('./helpers')
 
-function addDependencies (api) {
+function addDependencies (api, v3) {
   api.extendPackage({
     dependencies: {
-      vuetify: '^2.4.0',
+      vuetify: v3 ? '^3.0.0-alpha.0' : '^2.4.0',
     },
   })
 }
 
 function renderFiles (api, { opts }) {
-  const ext = api.hasPlugin('typescript') ? 'ts' : 'js'
+  const hasTS = api.hasPlugin('typescript')
+  const ext = hasTS ? 'ts' : 'js'
   const pluginFilename = `vuetify.${ext}`
 
   api.render({
-    [`./src/plugins/${pluginFilename}`]: '../templates/default/src/plugins/vuetify.js',
+    [`./src/plugins/${pluginFilename}`]: opts.useV3 ? '../templates/v3/src/plugins/vuetify.js' : '../templates/default/src/plugins/vuetify.js',
   }, {
     ...opts,
-    typescript: api.hasPlugin('typescript'),
+    typescript: hasTS,
   })
 
-  if (api.hasPlugin('typescript') && opts.useAlaCarte) {
+  if (hasTS && opts.useAlaCarte) {
     api.render({
       './src/shims-vuetify.d.ts': '../templates/default/src/shims-vuetify.d.ts',
     })
@@ -30,17 +30,31 @@ function renderFiles (api, { opts }) {
   // Render files if we're replacing
   const fs = require('fs')
   const routerPath = api.resolve(`./src/router.${ext}`)
+  const storePath = api.resolve(`./src/store.${ext}`)
   opts.router = fs.existsSync(routerPath)
+  opts.store = fs.existsSync(storePath)
 
+  let files = {
+    "./src/App.vue": `../templates/default/src/App.${ext}.vue`,
+    "./src/assets/logo.svg": "../templates/default/src/assets/logo.svg",
+    "./src/components/HelloWorld.vue": `../templates/default/src/components/HelloWorld.${ext}.vue`,
+  }
+
+  // replaceComponents is always true
   if (opts.replaceComponents) {
-    const files = {
-      './src/App.vue': `../templates/default/src/App.${ext}.vue`,
-      './src/assets/logo.svg': '../templates/default/src/assets/logo.svg',
-      './src/components/HelloWorld.vue': `../templates/default/src/components/HelloWorld.${ext}.vue`,
+    if (opts.useV3) {
+      files = {
+        "./src/App.vue": `../templates/v3/src/App.${ext}.vue`,
+        "./src/assets/logo.svg": "../templates/v3/src/assets/logo.svg",
+        "./src/components/HelloWorld.vue": `../templates/v3/src/components/HelloWorld.${ext}.vue`,
+        [api.entryFile]: "../templates/v3/src/main.js",
+      }
     }
 
     if (opts.router) {
-      files['./src/views/Home.vue'] = `../templates/default/src/views/Home.${ext}.vue`
+      files["./src/views/Home.vue"] = opts.v3
+        ? "../templates/v3/src/views/Home.js.vue"
+        : `../templates/default/src/views/Home.${ext}.vue`
     }
 
     api.render(files, opts)
@@ -48,8 +62,8 @@ function renderFiles (api, { opts }) {
 }
 
 function addImports (api) {
-  api.injectImports(api.entryFile, 'import vuetify from \'./plugins/vuetify\';')
-  api.injectRootOptions(api.entryFile, 'vuetify')
+  api.injectImports(api.entryFile, "import vuetify from './plugins/vuetify';")
+  api.injectRootOptions(api.entryFile, "vuetify")
 }
 
 function setHtmlLang (api, locale) {