+++ packages/vue-cli-plugin-vuetify-storybook/CHANGELOG.md
@@ -3,6 +3,15 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## <small>0.2.1 (2021-02-28)</small>
+
+* chore(eslintrc): remove vuetify config and update eslint settings ([a36555e](https://github.com/vuetifyjs/vue-cli-plugins/commit/a36555e))
+* fix(package.json): add missing dependencies and scripts for storybook (#173) ([d132b70](https://github.com/vuetifyjs/vue-cli-plugins/commit/d132b70)), closes [#173](https://github.com/vuetifyjs/vue-cli-plugins/issues/173) [#172](https://github.com/vuetifyjs/vue-cli-plugins/issues/172)
+
+
+
+
+
 # 0.2.0 (2020-02-05)